requirements.txt
pyyaml
pytest
boto3
azure-identity
azure-mgmt-compute
azure-mgmt-resource